Moved interfaces for Validator classes do Domain project, and implementtion of these to WebApi project. This to be able to use Validator classes in Query
