+++ src/Equinor.ProCoSys.Completion.Command/PunchItemCommands/DeletePunchItemAttachment/DeletePunchItemAttachmentCommandValidator.cs
@@ -2,7 +2,7 @@
 using System.Threading;
 using System.Threading.Tasks;
 using Equinor.ProCoSys.Completion.Command.Attachments;
-using Equinor.ProCoSys.Completion.Command.Validators.PunchItemValidators;
+using Equinor.ProCoSys.Completion.Domain.Validators;
 using FluentValidation;
 
 namespace Equinor.ProCoSys.Completion.Command.PunchItemCommands.DeletePunchItemAttachment;
