+++ src/Equinor.ProCoSys.Completion.Command/PunchItemCommands/UnclearPunchItem/UnclearPunchItemCommandValidator.cs
@@ -1,7 +1,7 @@
 ﻿using System;
 using System.Threading;
 using System.Threading.Tasks;
-using Equinor.ProCoSys.Completion.Command.Validators.PunchItemValidators;
+using Equinor.ProCoSys.Completion.Domain.Validators;
 using FluentValidation;
 
 namespace Equinor.ProCoSys.Completion.Command.PunchItemCommands.UnclearPunchItem;
