--- src/Equinor.ProCoSys.Completion.Command/ValidationException.cs
@@ -1,10 +1,0 @@
-﻿using System;
-
-namespace Equinor.ProCoSys.Completion.Command;
-
-public class ValidationException : Exception
-{
-    public ValidationException(string error) : base(error)
-    {
-    }
-}
