+++ src/Equinor.ProCoSys.Completion.Domain/Validators/ILibraryItemValidator.cs
@@ -3,7 +3,7 @@ using System.Threading;
 using System.Threading.Tasks;
 using Equinor.ProCoSys.Completion.Domain.AggregateModels.LibraryAggregate;
 
-namespace Equinor.ProCoSys.Completion.Command.Validators.LibraryItemValidators;
+namespace Equinor.ProCoSys.Completion.Domain.Validators;
 
 public interface ILibraryItemValidator
 {
