+++ src/Equinor.ProCoSys.Completion.Domain/Validators/IProjectValidator.cs
@@ -2,7 +2,7 @@
 using System.Threading;
 using System.Threading.Tasks;
 
-namespace Equinor.ProCoSys.Completion.Command.Validators.ProjectValidators;
+namespace Equinor.ProCoSys.Completion.Domain.Validators;
 
 public interface IProjectValidator
 {
