+++ src/Equinor.ProCoSys.Completion.WebApi/Validators/LibraryItemValidators/LibraryItemValidator.cs
@@ -4,9 +4,10 @@ using System.Threading;
 using System.Threading.Tasks;
 using Equinor.ProCoSys.Common;
 using Equinor.ProCoSys.Completion.Domain.AggregateModels.LibraryAggregate;
+using Equinor.ProCoSys.Completion.Domain.Validators;
 using Microsoft.EntityFrameworkCore;
 
-namespace Equinor.ProCoSys.Completion.Command.Validators.LibraryItemValidators;
+namespace Equinor.ProCoSys.Completion.WebApi.Validators.LibraryItemValidators;
 
 public class LibraryItemValidator : ILibraryItemValidator
 {
