+++ src/tests/Equinor.ProCoSys.Completion.Command.Tests/PunchItemCommands/DeletePunchItemLink/DeletePunchItemLinkCommandValidatorTests.cs
@@ -2,7 +2,7 @@
 using System.Threading.Tasks;
 using Equinor.ProCoSys.Completion.Command.Links;
 using Equinor.ProCoSys.Completion.Command.PunchItemCommands.DeletePunchItemLink;
-using Equinor.ProCoSys.Completion.Command.Validators.PunchItemValidators;
+using Equinor.ProCoSys.Completion.Domain.Validators;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
  using NSubstitute;
