+++ src/tests/Equinor.ProCoSys.Completion.Command.Tests/PunchItemCommands/UpdatePunchItem/UpdatePunchItemCommandValidatorTests.cs
@@ -1,7 +1,7 @@
 ﻿using System;
 using System.Threading.Tasks;
 using Equinor.ProCoSys.Completion.Command.PunchItemCommands.UpdatePunchItem;
-using Equinor.ProCoSys.Completion.Command.Validators.PunchItemValidators;
+using Equinor.ProCoSys.Completion.Domain.Validators;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
  using NSubstitute;
